Supress warning: narrowing conversion of 'rank' from 'int' to 'std::size_t'

+++ domain/test/integration_mpi/domain_nranks.cpp
@@ -239,7 +239,7 @@ void testAssignmentShift(const Box<double>& box)
     const bool useMixD  = (mixDBits.bx != maxTreeLevel<KeyType>{} || mixDBits.by != maxTreeLevel<KeyType>{} ||
                            mixDBits.bz != maxTreeLevel<KeyType>{});
 
-    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, rank};
+    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, static_cast<std::size_t>(rank)};
 
     std::vector<Real> x(coordinates.x().begin(), coordinates.x().end());
     std::vector<Real> y(coordinates.y().begin(), coordinates.y().end());
@@ -299,7 +299,7 @@ void removeParticle(const Box<double>& box)
     const bool useMixD  = (mixDBits.bx != maxTreeLevel<KeyType>{} || mixDBits.by != maxTreeLevel<KeyType>{} ||
                            mixDBits.bz != maxTreeLevel<KeyType>{});
 
-    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, rank};
+    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, static_cast<std::size_t>(rank)};
 
     std::vector<Real> x(coordinates.x().begin(), coordinates.x().end());
     std::vector<Real> y(coordinates.y().begin(), coordinates.y().end());
@@ -363,7 +363,7 @@ void testReapplySync(const Box<double>& box)
     const bool useMixD  = (mixDBits.bx != maxTreeLevel<KeyType>{} || mixDBits.by != maxTreeLevel<KeyType>{} ||
                            mixDBits.bz != maxTreeLevel<KeyType>{});
 
-    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, rank};
+    RandomCoordinates<Real, sfcKeyType<KeyType>> coordinates{numParticlesPerRank, box, static_cast<std::size_t>(rank)};
 
     std::vector<Real> x(coordinates.x().begin(), coordinates.x().end());
     std::vector<Real> y(coordinates.y().begin(), coordinates.y().end());
@@ -378,7 +378,7 @@ void testReapplySync(const Box<double>& box)
 
     // modify coordinates
     {
-        RandomCoordinates<Real, sfcKeyType<KeyType>> scord{domain.nParticles(), box, numRanks + rank};
+        RandomCoordinates<Real, sfcKeyType<KeyType>> scord{domain.nParticles(), box, static_cast<std::size_t>(numRanks + rank)};
         std::copy(scord.x().begin(), scord.x().end(), x.begin() + domain.startIndex());
         std::copy(scord.y().begin(), scord.y().end(), y.begin() + domain.startIndex());
         std::copy(scord.z().begin(), scord.z().end(), z.begin() + domain.startIndex());
